Rework to take the servicename as command-line argument
The aim is to allow for the storage of the config outside the program
tree. Also this would follow a more conventional program
vs. configuration separation.

An alternative would have been to just move the configuration-carrying
programs outside. The only difference is how to refer to the other
files, and whether to save 2-3 lines per config/program file; and to
have only one required file per configuration instead of one per
config*command, of course, since the combination of config and command
now happens by letting the user combine them on the command-line.